#72 - 중복코드 제거 및 가독성을 위해 require -> if로 변경

+++ src/main/kotlin/com/sat/board/application/CategoryCommandService.kt
@@ -54,7 +54,9 @@ class CategoryCommandService(
     fun delete(id: Long) {
         val category = getCategory(id)
         val exists = categoryRepository.existsByParentIdIs(id)
-        require(!exists) { throw ChildExistsException("자식이 존재하는 카테고리입니다. - $id") }
+        if (exists) {
+            throw ChildExistsException("자식이 존재하는 카테고리입니다. - $id")
+        }
         categoryRepository.delete(category)
     }
